namutrader/requirements.txt
requests>=2.31.0
websocket-client>=1.7.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
schedule>=1.2.0
# pandas-ta>=0.3.14b0  # Python 3.11 미지원 — RSI는 내장 순수 pandas 구현 사용
pytz>=2024.1
pytest>=7.4.0
